Refactor: directory cleanup
