--- hooks/use-fix-svg-transforms.ts
@@ -1,6 +1,0 @@
-/**
- * Re-export the useFixSvgTransforms hook from the lib/hooks directory
- * This file exists to maintain backward compatibility with imports from '@/hooks/use-fix-svg-transforms'
- */
-
-export { useFixSvgTransforms } from '@/lib/hooks/use-fix-svg-transforms';
